requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
decorator==5.1.1
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.1
idna==3.4
Pillow==9.2.0
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
self==2020.12.3
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
